Intermittent Push - see comments
Program will now take the program and build a class name,
identify the constructor and the parameter types for the constructor.

Need to  scan the program and identify if the parameter type is String
or int, and feed the String or int into the Constructor. Then create an
instance of the instruction and return it to the getInstruction method.